+++ app/src/main/java/nl/avans/drivioapp/service/DrivioApi.kt
@@ -1,0 +1,24 @@
+package nl.avans.drivioapp.service
+
+import com.squareup.moshi.Moshi
+import com.squareup.moshi.kotlin.reflect.KotlinJsonAdapterFactory
+import retrofit2.Retrofit
+import retrofit2.converter.moshi.MoshiConverterFactory
+
+object DrivioApi {
+
+    private val BASE_URL = "https://drivio.nl/api/v1/"
+
+    private val moshi = Moshi.Builder()
+        .add(KotlinJsonAdapterFactory())
+        .build()
+
+    private val retrofit = Retrofit.Builder()
+        .addConverterFactory(MoshiConverterFactory.create(moshi))
+        .baseUrl(BASE_URL)
+        .build()
+    
+    val advertisementService: AdvertisementService by lazy {
+        retrofit.create(AdvertisementService::class.java)
+    }
+}
