+++ app/src/main/java/nl/avans/drivioapp/viewModel/AdvertisementViewModel.kt
@@ -6,12 +6,14 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import kotlinx.coroutines.launch
 import nl.avans.drivioapp.model.Advertisement
+import nl.avans.drivioapp.repository.AdvertisementRepository
 import nl.avans.drivioapp.service.DrivioApi
 import retrofit2.Response
 
 private const val TAG = "AdvertisementViewModel"
 
 class AdvertisementViewModel : ViewModel() {
+    private val advertisementRepository = AdvertisementRepository()
     private val _getAdvertisementResponse: MutableLiveData<List<Advertisement>> = MutableLiveData()
     val getAdvertisementResponse: LiveData<List<Advertisement>>
         get() = _getAdvertisementResponse
@@ -27,14 +29,14 @@ class AdvertisementViewModel : ViewModel() {
 
     fun getAdvertisements() {
         viewModelScope.launch {
-            _getAdvertisementResponse.value = DrivioApi.retrofitService.getAdvertisements()
+            _getAdvertisementResponse.value = advertisementRepository.getAdvertisements()
         }
     }
 
 
     fun getAdvertisementById(advertisementId: Int) {
         viewModelScope.launch {
-            _getAdvertisementByIdResponse.value = DrivioApi.retrofitService.getAdvertisementById(advertisementId)
+            _getAdvertisementByIdResponse.value = advertisementRepository.getAdvertisementById(advertisementId)
         }
     }
 }
